+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support Rog Ally LED modes (basic)
 - Add on_lid_closed and on_external_power_changed events for running certain tasks
 - Anime dbus: add:
-  - SetOffWhenUnplugged
-  - SetOffWhenSuspended
-  - SetOffWhenLidClosed
+  - SetOffWhenUnplugged, also add asusctl CLI option
+  - SetOffWhenSuspended, also add asusctl CLI option
+  - SetOffWhenLidClosed, also add asusctl CLI option
 
 ### Changed
 - asusd: remove set_image_brightness for anime